refactor: (#21) Favorites 훅 분리
